Restructure charms client
Create new LocalCharmClient, which steals the method AddLocalCharm from
the original client.

This is because in future we will want to add extra dependencies to
AddLocalCharm, which are not relevant to the rest of the api methods.

(We are soon to add an S3-compatible endpoint to the controller for
uploading local charms, meaning we will need to inject an S3 client into
the api so that this Put operation can be performed)

Only a small number of the charm.Clients we construct actually need to
be able to call AddLocalCharm. Only two in production code (once in deploy,
once in refresh)

All non-deploy/refresh code (e.g. diffbundle, ssh_container,
charmresources) will just keep the reduced client, meaning they never
need to worry about these extra dependencies soon to be introduced
