+++ featuretests/client_macaroon_test.go
@@ -27,14 +27,14 @@ type clientMacaroonSuite struct {
 	apitesting.MacaroonSuite
 }
 
-func (s *clientMacaroonSuite) createTestClient(c *gc.C) *charms.Client {
+func (s *clientMacaroonSuite) createTestClient(c *gc.C) *charms.LocalCharmClient {
 	username := "testuser@somewhere"
 	s.AddModelUser(c, username)
 	s.AddControllerUser(c, username, permission.LoginAccess)
 	cookieJar := apitesting.NewClearableCookieJar()
 	s.DischargerLogin = func() string { return username }
 	api := s.OpenAPI(c, nil, cookieJar)
-	charmClient := charms.NewClient(api)
+	charmClient := charms.NewLocalCharmClient(api)
 
 	// Even though we've logged into the API, we want
 	// the tests below to exercise the discharging logic
@@ -56,7 +56,7 @@ func (s *clientMacaroonSuite) TestAddLocalCharmWithFailedDischarge(c *gc.C) {
 }
 
 func (s *clientMacaroonSuite) TestAddLocalCharmSuccess(c *gc.C) {
-	charmClient := charms.NewClient(s.APIState)
+	charmClient := charms.NewLocalCharmClient(s.APIState)
 	charmArchive := testcharms.Repo.CharmArchive(c.MkDir(), "dummy")
 	curl := charm.MustParseURL(
 		fmt.Sprintf("local:quantal/%s-%d", charmArchive.Meta().Name, charmArchive.Revision()),
